requirements.txt
# Core
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.10.*

# Database
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
psycopg2-binary==2.9.*
alembic==1.14.*

# Graph
neo4j==5.27.*

# Queue
celery[redis]==5.4.*
redis==5.2.*

# HTTP clients (for provider adapters)
httpx==0.28.*
aiohttp==3.11.*

# LLM
anthropic==0.43.*

# Utilities
python-dotenv==1.0.*
structlog==24.4.*
tenacity==9.0.*
